requirements.txt
flask
requests
beautifulsoup4
openai
serpapi
flask-limiter
python-dotenv